+++ Changelog.md
@@ -4,8 +4,10 @@
 
 - fix fence_gates and buttons not making sounds sometimes
 - add new block `wool_processor`
-    - can be used to convert wool blocks to other wool blocks
+    - can be used to convert wool blocks to the wool blocks added by this mod
+    - all previous recipes are no longer available
     - fixes recipe conflicts with some vanilla recipes
+    - support for JustEnoughItems, to show all the recipes
 
 ### Version 1.20.1 - 1.0